Create Sending your own data from the Arduino Yún to a Google Docs Spreadsheet
In this example we’ll demonstrate sending three types of data:

values from four analogue inputs (A0~A3)
the status of two digital inputs (D7 and D8)
the time and date from a DS3232 I2C real-time clock IC
With these types of data you should be able to represent all manner of things. We use the RTC as the time and date from it will match when the data was captured, not when the data was written to the spreadsheet. If you don’t have a DS3232 you can also use a DS1307.

If you’re not familiar with these parts and the required code please review this tutorial. When connecting your RTC – please note that SDA (data) is D2 and SCL (clock) is D3 on the Yún.